+++ .gitignore
@@ -1,4 +1,5 @@
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
+/dist
 
 # dependencies
 /node_modules
@@ -10,7 +11,6 @@
 
 # production
 /build
-/dist
 
 # misc
 .DS_Store
