+++ src/js/utils/loadPage.js
@@ -1,0 +1,6 @@
+const loadPage = (dex, tabObject) => {
+  dex.removeChild(dex.childNodes[0]);
+  dex.appendChild(tabObject);
+};
+
+export default loadPage;
