+++ src/js/utils/updateBook.js
@@ -1,0 +1,9 @@
+import AddNewBook from '../addNewBookApp';
+
+const updateBook = newBook => {
+  newBook.forEach(element => {
+    AddNewBook.updateNewBook(element);
+  });
+};
+
+export default updateBook;
